+++ Computational Geometry/Line.hpp
@@ -1,4 +1,4 @@
-#include "Vector.h"
+#include "Vector.hpp"
 
 namespace GeomCore{
     template <class coord_type, size_t dim = R3>
@@ -21,8 +21,8 @@ namespace GeomCore{
             void set_point(Vector<coord_type, dim>& _point);
     };
 
-    typedef Line<float, R2> Line2d;
-    typedef Line<float, R3> Line3d;
+    typedef Line<double, R2> Line2d;
+    typedef Line<double, R3> Line3d;
     
     template <class coord_type, size_t dim>
     inline Vector<coord_type, dim> Line<coord_type, dim>::get_point() const{
